pyproject.toml
[project]
name = "listener"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.121.0",
    "jinja2>=3.1.6",
    "langchain>=1.0.4",
    "langchain-core>=1.0.3",
    "langchain-google-genai>=3.0.1",
    "pika>=1.3.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "ruff>=0.14.4",
    "torch>=2.9.0",
    "transformers>=4.57.1",
    "uvicorn>=0.38.0",
]
